feat: Add sorting functionality to order-page component
This commit adds sorting functionality to the order-page component. Three buttons are added to the toolbar, allowing the user to sort the heroes table by name, canFly, or color. The corresponding sortByPipe is applied to the heroes array to sort the table based on the selected property. This enhancement improves the user experience by providing the ability to easily sort the heroes table in different ways.